moved TitleBarIcon from Controls to Utils
