+++ App/Views/DirectoryPickerWindow.xaml.cs
@@ -8,7 +8,7 @@ using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Media;
 using WinRT.Interop;
 using WinUIEx;
-using Coder.Desktop.App.Controls;
+using Coder.Desktop.App.Utils;
 
 namespace Coder.Desktop.App.Views;
