+++ App/Views/FileSyncListWindow.xaml.cs
@@ -2,7 +2,7 @@ using Coder.Desktop.App.ViewModels;
 using Coder.Desktop.App.Views.Pages;
 using Microsoft.UI.Xaml.Media;
 using WinUIEx;
-using Coder.Desktop.App.Controls;
+using Coder.Desktop.App.Utils;
 
 namespace Coder.Desktop.App.Views;
