+++ App/Views/SignInWindow.xaml.cs
@@ -6,6 +6,7 @@ using Coder.Desktop.App.Views.Pages;
 using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Media;
+using Coder.Desktop.App.Utils;
 
 namespace Coder.Desktop.App.Views;
